html/cross-origin-embedder-policy/credentialless/cors-or-credentialless/dedicated-worker.tentative.https.html

<!doctype html>
<title>CORS or Credentialless and dedicated worker</title>
<meta name="timeout" content="long">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/dispatcher.js"></script>
<script>

const same_origin = get_host_info().HTTPS_ORIGIN;
const cross_origin = get_host_info().HTTPS_REMOTE_ORIGIN;
const cookie_key = "credentialless_dedicated_worker";
const cookie_same_origin = "same_origin";
const cookie_cross_origin = "cross_origin";

promise_test(async test => {

  await Promise.all([
    setCookie(same_origin, cookie_key, cookie_same_origin),
    setCookie(cross_origin, cookie_key, cookie_cross_origin),
  ]);

  // One window with COEP:none. (control)
  const w_control_token = token();
  const w_control_url = same_origin + executor_path +
    coep_none + `&uuid=${w_control_token}`
  const w_control = window.open(w_control_url);
  add_completion_callback(() => w_control.close());

  // One window with COEP:credentialless. (experiment)
  const w_credentialless_token = token();
  const w_credentialless_url = same_origin + executor_path +
    coep_credentialless + `&uuid=${w_credentialless_token}`;
  const w_credentialless = window.open(w_credentialless_url);
  add_completion_callback(() => w_credentialless.close());

  const dedicatedWorkerTest = function(
    description, origin, mode, credentials, workerWithCredentialless, workerOptions,
    expected_cookies_control,
    expected_cookies_credentialless)
  {
    promise_test_parallel(async t => {
      const token_1 = token();
      const token_2 = token();

      const w_worker_src_1 = same_origin + executor_js_path +
        workerWithCredentialless + `&uuid=${token_1}`;
      const w_post_1 = JSON.stringify({url: showRequestHeaders(origin, token_1),
        mode: mode, credentials: credentials});
      send(w_control_token, `
        const worker = new Worker("${w_worker_src_1}", ${JSON.stringify(workerOptions)});
        worker.postMessage(${w_post_1});
      `);

      const w_worker_src_2 = same_origin + executor_js_path +
        workerWithCredentialless + `&uuid=${token_2}`;
      const w_post_2 = JSON.stringify({url: showRequestHeaders(origin, token_2),
        mode: mode, credentials: credentials});
      send(w_credentialless_token, `
        const worker = new Worker("${w_worker_src_2}", ${JSON.stringify(workerOptions)});
        worker.postMessage(${w_post_2})
      `);

      const headers_control = JSON.parse(await receive(token_1));
      const headers_credentialless = JSON.parse(await receive(token_2));

      assert_equals(parseCookies(headers_control)[cookie_key],
        expected_cookies_control,
        "coep:none => ");
      assert_equals(parseCookies(headers_credentialless)[cookie_key],
        expected_cookies_credentialless,
        "coep:credentialless => ");
    }, `fetch ${description}`)
  };

  // Cookies are never sent with credentials='omit'
  dedicatedWorkerTest("same-origin + no-cors + credentials:omit + credentialless worker",
    same_origin, 'no-cors', 'omit', coep_credentialless, {},
    undefined,
    undefined);
  dedicatedWorkerTest("cross-origin + no-cors + credentials:omit + credentialless worker",
    cross_origin, 'no-cors', 'omit', coep_credentialless, {},
    undefined,
    undefined);
  dedicatedWorkerTest("same-origin + no-cors + credentials:omit + module worker + credentialless worker",
    same_origin, 'no-cors', 'omit', coep_credentialless, {type: 'module'},
    undefined,
    undefined);
  dedicatedWorkerTest("same-origin + no-cors + credentials:omit",
    same_origin, 'no-cors', 'omit', coep_none, {},
    undefined,
    undefined);
  dedicatedWorkerTest("cross-origin + no-cors + credentials:omit",
    cross_origin, 'no-cors', 'omit', coep_none, {},
    undefined,
    undefined);
  dedicatedWorkerTest("same-origin + no-cors + credentials:omit + module worker",
    same_origin, 'no-cors', 'omit', coep_none, {type: 'module'},
    undefined,
    undefined);

  // Same-origin request contains Cookies.
  dedicatedWorkerTest("same-origin + no-cors + credentials:include + credentialless worker",
    same_origin, 'no-cors', 'include', coep_credentialless, {},
    cookie_same_origin,
    cookie_same_origin);
  dedicatedWorkerTest("same-origin + no-cors + credentials:same-origin + credentialless worker",
    same_origin, 'no-cors', 'same-origin', coep_credentialless, {},
    cookie_same_origin,
    cookie_same_origin);

  // Cross-origin CORS requests contains Cookies, if credentials mode is set to
  // 'include'. This does not depends on COEP.
  dedicatedWorkerTest("cross-origin + cors + credentials:include + credentialless worker",
    cross_origin, 'cors', 'include', coep_none, {},
    cookie_cross_origin,
    cookie_cross_origin);
  dedicatedWorkerTest("cross-origin + cors + same-origin-credentials + credentialless worker",
    cross_origin, 'cors', 'same-origin', coep_none, {},
    undefined,
    undefined);

  // Cross-origin no-CORS requests includes Cookies when:
  // 1. credentials mode is 'include'
  // 2. COEP: is not credentialless.
  dedicatedWorkerTest("cross-origin + no-cors + credentials:include",
    cross_origin, 'no-cors', 'include', coep_none, {},
    cookie_cross_origin,
    undefined);

  dedicatedWorkerTest("cross-origin + no-cors + credentials:same-origin",
    cross_origin, 'no-cors', 'same-origin', coep_none, {},
    undefined,
    undefined);
})


</script>
